bugfix & tests
