+++ .gitignore
@@ -7,4 +7,5 @@ node_modules/
 octomind.config.json
 octomind-cli-debug/
 openapi.yaml
+vhs/
